Clean up CA generation command code
Add hash algo selection method
